Deleted demoDrafts in utils and moved methods in helpers. Remove UK short demo from surveys which should not be saved
