+++ src/screens/utils/helpers.js
@@ -63,3 +63,11 @@ export const prepareDraftForSubmit = (draft, survey) => {
 
   return result
 }
+
+const demoDrafts = [
+  { id: 19, title: 'Stoplight Demo - Spanish' },
+  { id: 20, title: 'Stoplight Demo - English' }
+]
+
+export const isDemoDraft = ({ id, title }) =>
+  demoDrafts.some(draft => draft.id === id && draft.title === title)
